+++ scheduler-service/src/main/java/com/codeaddi/scheduler_service/controller/SessionController.java
@@ -3,7 +3,7 @@ package com.codeaddi.scheduler_service.controller;
 import com.codeaddi.scheduler_service.controller.db.SessionsService;
 import com.codeaddi.scheduler_service.model.http.StandardResponse;
 import com.codeaddi.scheduler_service.model.http.enums.Status;
-import com.codeaddi.scheduler_service.model.repository.Session;
+import com.codeaddi.scheduler_service.model.repository.sessions.Session;
 import jakarta.persistence.EntityNotFoundException;
 import java.util.List;
 import lombok.extern.slf4j.Slf4j;
