+++ scheduler-service/src/main/java/com/codeaddi/scheduler_service/controller/db/SessionsService.java
@@ -1,7 +1,7 @@
 package com.codeaddi.scheduler_service.controller.db;
 
-import com.codeaddi.scheduler_service.model.repository.Session;
-import com.codeaddi.scheduler_service.model.repository.SessionRepository;
+import com.codeaddi.scheduler_service.model.repository.sessions.Session;
+import com.codeaddi.scheduler_service.model.repository.sessions.SessionRepository;
 import jakarta.persistence.EntityNotFoundException;
 import java.util.List;
 import lombok.extern.slf4j.Slf4j;
