+++ scheduler-service/src/test/java/com/codeaddi/scheduler_service/controller/db/SessionsServiceTests.java
@@ -5,8 +5,8 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.Mockito.*;
 
-import com.codeaddi.scheduler_service.model.repository.Session;
-import com.codeaddi.scheduler_service.model.repository.SessionRepository;
+import com.codeaddi.scheduler_service.model.repository.sessions.Session;
+import com.codeaddi.scheduler_service.model.repository.sessions.SessionRepository;
 import com.codeaddi.scheduler_service.testUtils.TestData;
 import jakarta.persistence.EntityNotFoundException;
 import java.util.Optional;
